+++ crates/near-contract-transport/src/views.rs
@@ -1,11 +1,11 @@
 mod args;
-mod call;
+mod deserialize;
 mod errors;
 mod result;
 mod traits;
 
 pub use args::ViewArgs;
-pub use call::{DeserializationError, Deserializer, borsh_de, json_de};
+pub use deserialize::{DeserializationError, Deserializer, borsh_de, json_de};
 pub use errors::TransportError;
 pub use result::{ObservedState, SerializedObservation};
 pub use traits::{HasPollInterval, PollInterval, ViewContract, WatchContractState};
