+++ crates/near-contract-transport/src/views/deserialize.rs
